float

+++ dagrad/hfunction/h_functions.py
@@ -34,7 +34,7 @@ class SCCPowerIteration(nn.Module):
         return self._dummy_param.device
 
     def initialize_eigenvectors(self, adj_mtx):
-        self.v, self.vt = torch.ones(size=(2, self.d), device=self.device, dtype=torch.double)
+        self.v, self.vt = torch.ones(size=(2, self.d), device=self.device, dtype=torch.float)
         self.v = normalize(self.v)
         self.vt = normalize(self.vt)
         return self.power_iteration(adj_mtx, 5)
@@ -85,7 +85,7 @@ class SCCPowerIteration(nn.Module):
         # matrix = self.power_iteration(4)
         matrix = self.initialize_eigenvectors(adj_mtx)
 
-        gradient = torch.zeros(size=(self.d, self.d), device=self.device, dtype=torch.double)
+        gradient = torch.zeros(size=(self.d, self.d), device=self.device, dtype=torch.float)
         for scc in self.scc_list:
             if len(scc) == self.d:
                 v = self.v
